# API/src/main/java/com/kbaje/eshop/dto/CreateProductDto.java
package com.kbaje.eshop.dto;

import java.math.BigDecimal;

public class CreateProductDto {

    public String name;
    public String description;
    public String imageUrl;
    public BigDecimal price;

}